Add files via upload
Still learning and with great passion in ethical hacking/cybersecurity. This is one of the successful projects I created with the help of AI, YouTube and the internet. It is a keylogger that uses email, discord and USB. All this is tested. It is encrypted. Teaches one to know how hackers operate and how one would fall for an attack. Still progressing, in a year I will have enough knowledge and skills.